fix: Cast avg_completion_seconds to float for MySQL compatibility
- MySQL returns TIMESTAMPDIFF results as strings, not floats
- Added explicit (float) cast before round() to prevent TypeError
- Fixed in both MultiCustodianTransferService and SettlementService

+++ app/Domain/Custodian/Services/MultiCustodianTransferService.php
@@ -530,7 +530,7 @@ class MultiCustodianTransferService
                 'external' => $stats->external ?? 0,
                 'bridge' => $stats->bridge ?? 0,
             ],
-            'avg_completion_seconds' => round($stats->avg_completion_seconds ?? 0, 2),
+            'avg_completion_seconds' => round((float) ($stats->avg_completion_seconds ?? 0), 2),
             'success_rate' => $stats->total_transfers > 0 
                 ? round(($stats->completed / $stats->total_transfers) * 100, 2) 
                 : 0,
